Enhance game mechanics with new enemy types and tower features
- Added armored, fast, and boss mobs with unique abilities and unit tests.
- Implemented sniper and rapid-fire towers with distinct mechanics and verified unit tests.
- Updated balancing documentation to reflect new mob types and their behaviors.
- Refactored game entity methods for improved accuracy and performance.
- Enhanced test coverage for new mob behaviors and tower functionalities.

+++ v1/internal/game/entity.go
@@ -9,7 +9,7 @@ type Entity interface {
 	Frame() *ebiten.Image              // Get the current image frame of the entity (static or animated)
 	Update(dt float64) error           // Update the entity state using delta time
 	Draw(screen *ebiten.Image)         // Draw the entity to the screen
-	Position() (x, y int)              // Get the current position of the entity
+	Position() (x, y float64)          // Get the current position of the entity
 	Bounds() (x, y, width, height int) // Get the bounding box of the entity
 	Hitbox() (x, y, width, height int) // Get the hitbox of the entity
 	Static() bool                      // Check if the entity is static (does not move)
